+++ CHANGELOG.md
@@ -35,6 +35,7 @@ v5.10.0
 - Fix memory leak with Scotch graph initialization (!2518)
 - Fix memory leak with GS in AssemblyMapDG initialization (!2519)
 - Use static linkage for FFTW to avoid problem with MKL (!2474)
+- Fix unitialized variables in Collection and DriverParareal (!2521)
 
 **IncNavierStokesSolver**
 - Added an option to process Reynolds Stress fields at a higher polynomial order (!2303)